requirements.txt
fastapi
uvicorn
websockets
PyYAML
